Created the managepatientpage which can edit patients and simplified the patientselect page
