+++ src/pages/PatientSelect.js
@@ -9,12 +9,6 @@ import PatientAddButton from  "../components/HomePages/PatientAddButton";
 function PatientSelect(){
 	const [buttonPopup, setButtonPopup] = useState(false);
     const [patientsData, setPatientsData] = useState([]);
-	const [updatedPatientName, setUpdatedPatientName] = useState("");
-	const [updatedPatientAge, setUpdatedPatientAge] = useState("");
-	const [updatedPatientGender, setUpdatedPatientGender] = useState("");
-	const [updatedPatientBedNO, setUpdatedPatientBedNO] = useState("");
-	const [dataIdToBeUpdated, setDataIdToBeUpdated] = useState("");
-
 
     useEffect(() => {
         db.collection("patientsData").onSnapshot((snapshot) => {
@@ -28,75 +22,18 @@ function PatientSelect(){
         }, 
     []);
 
-
-	const updateData = (e) => {
-		e.preventDefault();
-		db.collection("patientsData").doc(dataIdToBeUpdated).update({
-		name: updatedPatientName,
-		age: updatedPatientAge,
-		gender: updatedPatientGender,
-		bedNO: updatedPatientBedNO
-		});
-	
-		setUpdatedPatientAge("");
-		setUpdatedPatientName("");
-		setUpdatedPatientGender("");
-		setUpdatedPatientBedNO("");
-		setDataIdToBeUpdated("");
-	};
-
-
-	const deleteData = (id) => {
-        db.collection("patientsData").doc(id).delete(); }
-
-	return (<>
-
-	<div className="App">
-	{!dataIdToBeUpdated ? (
-		
-		<div className="App__buttons">
-		<PatientAddButton/>
-		</div>
-		) : (
-		<div className="App__Updateform">
-		<input
-			type="text"
-			placeholder="Name"
-			value={updatedPatientName}
-			onChange={(e) => setUpdatedPatientName(e.target.value)}
-		/>
-		<input
-			type="text"
-			placeholder="Age"
-			value={updatedPatientAge}
-			onChange={(e) => setUpdatedPatientAge(e.target.value)}
-		/>
-		<input
-			type="text"
-			placeholder="Gender"
-			value={updatedPatientGender}
-			onChange={(e) => setUpdatedPatientGender(e.target.value)}
-		/>
-				<input
-			type="text"
-			placeholder="Bed Number"
-			value={updatedPatientBedNO}
-			onChange={(e) => setUpdatedPatientBedNO(e.target.value)}
-		/>
-		<button onClick={updateData}>Update</button>
-		</div>
-		)}
-
+return(<>
 
 	  <div className="App__DataDisplay">
+	  <div className="App__buttons">
+		<PatientAddButton/>
+		</div>
 		<table>
 			<tr>
 			<th>NAME</th>
 			<th>AGE</th>
 			<th>GENDER</th>
 			<th>BED NUMBER</th>
-			<th>Update</th>
-			<th>Delete</th>
 			<th>View Details</th>
 			<th>Select</th>
 			</tr>
@@ -107,32 +44,7 @@ function PatientSelect(){
 			<td>{data.age}</td>
 			<td>{data.gender}</td>
 			<td>{data.bedNO}</td>
-
 			<td>
-				<button
-				onClick={() => {
-                    setDataIdToBeUpdated(id);
-                    setUpdatedPatientAge(data.age);
-                    setUpdatedPatientName(data.name);
-                    setUpdatedPatientGender(data.gender);
-                    setUpdatedPatientBedNO(data.bedNO);
-					//<PatientAddButton/>
-				}}
-				>
-				Update
-				</button>
-			</td>
-			<td>
-				<button
-				onClick={() => {
-					deleteData(id);
-				}}
-				>
-				Delete
-				</button>
-			</td>
-			<td>
-			<>
 			<button onClick={() => 
 				 {setButtonPopup(true)}
 				 }> View Details </button>
@@ -147,13 +59,12 @@ function PatientSelect(){
 					
 					
 				</Popup>
-			</>
+			
 			</td>
 			</tr>
 		))}
 		
 		</table>
-	</div>
 	</div>
 	</>
 );
